+++ database/migrations/2020_05_05_051947_create_versions_table.php
@@ -24,7 +24,7 @@ class CreateVersionsTable extends Migration
             $table->string('mr_audio_file')->nullable();
             $table->string('voice_audio_file')->nullable();
             $table->text('lyrics')->nullable();
-            $table->unsignedTinyInteger('is_opened')->default(0); //0:전체 공개 1:회원공개 2:팔로워공개 3:협업자공개 4:비공개
+            $table->unsignedInteger('is_opened')->default(0); //0:전체 공개 1:회원공개 2:팔로워공개 3:협업자공개 4:비공개
             $table->timestamps();
             $table->softDeletes();
